feat: use models
