+++ apps/www/src/api/contract.ts
@@ -1,16 +1,14 @@
-import {
-  healthCheckSchemaResponses,
-  fetchGpgFromUpstreamSchemaResponses,
-  fetchTextFromUpstreamSchemaResponses,
-  purgeViewWindowSchemaResponses,
-} from "./schemas/responses";
-import {
-  fetchTextFromUpstreamQuerySchemaDto,
-  purgeViewWindowHeadersSchemaDto,
-} from "./schemas/dtos";
 import { c } from "../ts-rest/root";
 import type { Keys } from "ts-roids";
 import { createContract } from "~/@ashgw/ts-rest";
+import {
+  healthCheckSchemaResponses,
+  fetchGpgFromUpstreamSchemaResponses,
+  fetchTextFromUpstreamQuerySchemaDto,
+  purgeViewWindowHeadersSchemaDto,
+  fetchTextFromUpstreamSchemaResponses,
+  purgeViewWindowSchemaResponses,
+} from "~/api/models";
 
 // TODO: add summary and shit here so AI can use it
 export const contract = createContract(c)({
