+++ apps/www/src/api/models/_shared/responses.ts
@@ -1,0 +1,27 @@
+import { c } from "~/ts-rest/root";
+import { createSchemaResponses, httpErrorSchema } from "~/@ashgw/ts-rest";
+
+// ========== Schemas ==========
+
+export const okSchemaResponse = createSchemaResponses({
+  200: c.noBody(),
+});
+
+export const cronAuthedMiddlewareSchemaResponse = createSchemaResponses({
+  401: httpErrorSchema
+    .unauthorized()
+    .describe("Missing or invalid x-cron-token"),
+});
+
+export const rateLimiterMiddlwareSchemaResponse = createSchemaResponses({
+  429: httpErrorSchema
+    .tooManyRequests()
+    .describe("Exceeded the allowed window to make requests"),
+});
+
+export const internalErrorSchemaResponse = createSchemaResponses({
+  500: httpErrorSchema.internal().describe("Internal failure"),
+});
+
+// ========== Types ==========
+// no needed for these
