+++ apps/www/src/api/models/fetchTextFromUpstream/dtos.ts
@@ -1,0 +1,17 @@
+import { z } from "zod";
+
+export const fetchTextFromUpstreamQuerySchemaDto = z
+  .object({
+    revalidateSeconds: z
+      .string()
+      .regex(/^\d+$/)
+      .transform((v) => Number(v))
+      .pipe(z.number().int().min(60).max(86400))
+      .optional()
+      .describe("Override ISR revalidate in seconds, 60..86400"),
+  })
+  .passthrough();
+
+export type FetchTextFromUpstreamQueryDto = z.infer<
+  typeof fetchTextFromUpstreamQuerySchemaDto
+>;
