+++ apps/www/src/api/models/fetchTextFromUpstream/index.ts
@@ -1,0 +1,10 @@
+export type { FetchTextFromUpstreamQueryDto } from "./dtos";
+export { fetchTextFromUpstreamQuerySchemaDto } from "./dtos";
+export {
+  fetchGpgFromUpstreamSchemaResponses,
+  fetchTextFromUpstreamSchemaResponses,
+} from "./responses";
+export type {
+  FetchGpgFromUpstreamResponses,
+  FetchTextFromUpstreamResponses,
+} from "./responses";
