+++ apps/www/src/api/models/healthCheck/dtos.ts
@@ -1,0 +1,1 @@
+// not needed
