+++ apps/www/src/api/models/healthCheck/index.ts
@@ -1,0 +1,2 @@
+export { healthCheckSchemaResponses } from "./responses";
+export type { HealthCheckResponses } from "./responses";
