+++ apps/www/src/api/models/webhooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./purgeViewWindow";
