+++ apps/www/src/api/models/webhooks/purgeViewWindow/index.ts
@@ -1,0 +1,4 @@
+export { purgeViewWindowHeadersSchemaDto } from "./dtos";
+export type { PurgeViewWindowHeadersDto } from "./dtos";
+export { purgeViewWindowSchemaResponses } from "./responses";
+export type { PurgeViewWindowResponses } from "./responses";
