+++ apps/www/src/api/models/webhooks/purgeViewWindow/responses.ts
@@ -1,0 +1,19 @@
+import type { InferResponses } from "~/@ashgw/ts-rest";
+import { createSchemaResponses } from "~/@ashgw/ts-rest";
+import {
+  cronAuthedMiddlewareSchemaResponse,
+  internalErrorSchemaResponse,
+  okSchemaResponse,
+  rateLimiterMiddlwareSchemaResponse,
+} from "../../_shared/responses";
+
+export const purgeViewWindowSchemaResponses = createSchemaResponses({
+  ...rateLimiterMiddlwareSchemaResponse,
+  ...cronAuthedMiddlewareSchemaResponse,
+  ...okSchemaResponse,
+  ...internalErrorSchemaResponse,
+});
+
+export type PurgeViewWindowResponses = InferResponses<
+  typeof purgeViewWindowSchemaResponses
+>;
